Don't copy progress tool for root user (#2897)
This way we don't have to create cache directories under a different user to launch the progress tool. Longer term I want to never copy the tool.

+++ Sparkle/InstallerProgress/InstallerProgressAppController.m
@@ -62,13 +62,14 @@ static const NSTimeInterval SUTerminationTimeDelay = 0.3;
     BOOL _submittedLauncherJob;
     BOOL _willTerminate;
     BOOL _applicationInitiallyAlive;
+    BOOL _copiedApplication;
 }
 
 - (instancetype)initWithApplication:(NSApplication *)application arguments:(NSArray<NSString *> *)arguments delegate:(id<InstallerProgressDelegate>)delegate
 {
     self = [super init];
     if (self != nil) {
-        if (arguments.count != 3) {
+        if (arguments.count != 4) {
             SULog(SULogLevelError, @"Error: Invalid number of arguments supplied: %@", arguments);
             [self cleanupAndExitWithStatus:EXIT_FAILURE error:nil];
         }
@@ -114,7 +115,9 @@ static const NSTimeInterval SUTerminationTimeDelay = 0.3;
         _application = application;
         _delegate = delegate;
         
-        [delegate loadLocalizationStringsFromHost:host];
+        BOOL copiedApplication = arguments[3].boolValue;
+        [delegate loadLocalizationStringsFromHost:host copiedApplication:copiedApplication];
+        _copiedApplication = copiedApplication;
         
         _connection.exportedInterface = [NSXPCInterface interfaceWithProtocol:@protocol(SPUInstallerAgentProtocol)];
         _connection.exportedObject = self;
@@ -193,21 +196,23 @@ static const NSTimeInterval SUTerminationTimeDelay = 0.3;
     [_statusInfo invalidate];
     [_connection invalidate];
     
-    // Remove the agent bundle; it is assumed this bundle is in a temporary/cache/support directory
-    NSError *theError = nil;
-    NSString *bundlePath = [[NSBundle mainBundle] bundlePath];
-    
-    if (![[NSFileManager defaultManager] removeItemAtPath:bundlePath error:&theError]) {
-        SULog(SULogLevelError, @"Couldn't remove agent bundle: %@.", theError);
-    } else {
-        // There should be nothing else in the parent temporary directory given to us,
-        // so let us try to remove it. Note rmdir() will fail if there are unexpectably other
-        // items present
-        NSString *parentDirectory = bundlePath.stringByDeletingLastPathComponent;
-        const char *fileSystemRepresentation = parentDirectory.fileSystemRepresentation;
-        if (fileSystemRepresentation != NULL) {
-            if (rmdir(fileSystemRepresentation) != 0) {
-                SULog(SULogLevelError, @"Failed to remove parent agent bundle directory: %@: %d", parentDirectory, errno);
+    if (_copiedApplication) {
+        // Remove the agent bundle; it is assumed this bundle is in a temporary/cache/support directory
+        NSError *theError = nil;
+        NSString *bundlePath = [[NSBundle mainBundle] bundlePath];
+        
+        if (![[NSFileManager defaultManager] removeItemAtPath:bundlePath error:&theError]) {
+            SULog(SULogLevelError, @"Couldn't remove agent bundle: %@.", theError);
+        } else {
+            // There should be nothing else in the parent temporary directory given to us,
+            // so let us try to remove it. Note rmdir() will fail if there are unexpectably other
+            // items present
+            NSString *parentDirectory = bundlePath.stringByDeletingLastPathComponent;
+            const char *fileSystemRepresentation = parentDirectory.fileSystemRepresentation;
+            if (fileSystemRepresentation != NULL) {
+                if (rmdir(fileSystemRepresentation) != 0) {
+                    SULog(SULogLevelError, @"Failed to remove parent agent bundle directory: %@: %d", parentDirectory, errno);
+                }
             }
         }
     }
